+++ failgood/src/test/kotlin/failgood/junit/it/fixtures/packagewith1test/TestFixture.kt
@@ -3,6 +3,6 @@ package failgood.junit.it.fixtures.packagewith1test
 import failgood.describe
 
 object TestFixture {
-    val CONTEXT_NAME = "the test fixture"
+    const val CONTEXT_NAME = "the test fixture"
     val context = describe(CONTEXT_NAME) {}
 }